backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.7
langchain-openai==0.2.6
openai==1.54.0
pydantic-settings==2.6.1
python-multipart==0.0.12
easyocr==1.7.2
Pillow==11.0.0
numpy==1.26.4
pytest==8.3.3
httpx==0.27.2
